Fix misuse of SimpleDateFormat
One does not simply use the Java date APIs... Bizarre timezones such as
BDT were being generated, because the SimpleDateFormat instance was
not being reset to the correct timezone after each use. After parsing
a BDT date, the SDF's calendar would maintain the BDT timezone, leading
to it using it for future format() operations.

Also SDF is not thread safe, so an instance per thread (or
synchronisation) is required.