Use URI resolution instead of templating
UriComponentsBuilder is an abomination. The need to explicitly call
encode() is ridiculous. Using a template like `http://foo.com{path}`
with the template value for `path` being `/abc` results in the host
being `foo.com/abc`. So you need knowledge of the template string to
correctly interpolate values (e.g. to strip a leading / from the path),
which defeats the point of using a URI template in the first place.